requirements.txt
fastapi
uvicorn
python-jose
apscheduler
pydantic<2.0.0
python-multipart
sqlalchemy
pymysql
cryptography
python-dotenv
passlib[bcrypt]